Refactor: Standardize ready endpoint for Office Service
Makes the ready endpoint for the Office Service consistent with other services by changing it from `/health/ready` to `/ready`.

- Moved `ready_check` function from `office_service/api/health.py` to `office_service/app/main.py`.
- Updated the route decorator to `@app.get("/ready")`.
- Ensured necessary imports are in place in `main.py`.
