+++ services/office_service/app/main.py
@@ -15,6 +15,7 @@ from core.exceptions import (
     ValidationError,
 )
 from core.logging_config import setup_logging
+from datetime import datetime
 from fastapi import FastAPI, Request
 from fastapi.responses import JSONResponse
 from schemas import ApiError
@@ -303,3 +304,16 @@ async def read_root():
     """Hello World root endpoint"""
     logger.info("Root endpoint accessed")
     return {"message": "Hello World", "service": "Office Service"}
+
+
+@app.get("/ready")
+async def ready_check():
+    """
+    Simple readiness check.
+    """
+    return {
+        "status": "ok",
+        "service": settings.APP_NAME,
+        "version": settings.APP_VERSION,
+        "timestamp": datetime.utcnow().isoformat(),
+    }
